requirements.txt
streamlit
boto3
Pillow
